refactor : non chain , chain filter 추상화
