+++ src/main/java/nextstep/auth/exception/RoleAuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.auth.secured;
+package nextstep.auth.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
