+++ src/main/java/nextstep/auth/interceptor/TokenAuthenticationInterceptor.java
@@ -1,5 +1,6 @@
 package nextstep.auth.interceptor;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.auth.exception.AuthenticationException;
 import nextstep.auth.token.JwtTokenProvider;
@@ -12,9 +13,10 @@ import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.util.stream.Collectors;
 
-public class TokenAuthenticationInterceptor implements HandlerInterceptor {
+public class TokenAuthenticationInterceptor extends AuthenticationNonChainHandler {
     private LoginMemberService loginMemberService;
     private JwtTokenProvider jwtTokenProvider;
 
@@ -24,31 +26,45 @@ public class TokenAuthenticationInterceptor implements HandlerInterceptor {
     }
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        String content = request.getReader().lines().collect(Collectors.joining(System.lineSeparator()));
-        TokenRequest tokenRequest = new ObjectMapper().readValue(content, TokenRequest.class);
+    protected LoginMember createAuthentication(HttpServletRequest request) {
+        try {
+            String content = request.getReader().lines().collect(Collectors.joining(System.lineSeparator()));
 
-        String principal = tokenRequest.getEmail();
-        String credentials = tokenRequest.getPassword();
+            TokenRequest tokenRequest = new ObjectMapper().readValue(content, TokenRequest.class);
 
-        LoginMember loginMember = loginMemberService.loadUserByUsername(principal);
+            String principal = tokenRequest.getEmail();
+            String credentials = tokenRequest.getPassword();
 
-        if (loginMember == null) {
+            LoginMember loginMember = loginMemberService.loadUserByUsername(principal);
+
+            if (loginMember == null) {
+                throw new AuthenticationException();
+            }
+
+            if (!loginMember.checkPassword(credentials)) {
+                throw new AuthenticationException();
+            }
+
+            return loginMember;
+        } catch (Exception e) {
             throw new AuthenticationException();
         }
+    }
 
-        if (!loginMember.checkPassword(credentials)) {
-            throw new AuthenticationException();
-        }
-
+    @Override
+    protected void afterHandle(LoginMember loginMember, HttpServletResponse response) {
         String token = jwtTokenProvider.createToken(loginMember.getEmail(), loginMember.getAuthorities());
         TokenResponse tokenResponse = new TokenResponse(token);
 
-        String responseToClient = new ObjectMapper().writeValueAsString(tokenResponse);
-        response.setStatus(HttpServletResponse.SC_OK);
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.getOutputStream().print(responseToClient);
+        String responseToClient = null;
+        try {
+            responseToClient = new ObjectMapper().writeValueAsString(tokenResponse);
+            response.setStatus(HttpServletResponse.SC_OK);
+            response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+            response.getOutputStream().print(responseToClient);
+        } catch (Exception e) {
+            throw new AuthenticationException();
+        }
 
-        return false;
     }
 }
