+++ src/main/java/nextstep/auth/interceptor/UsernamePasswordAuthenticationFilter.java
@@ -1,0 +1,50 @@
+package nextstep.auth.interceptor;
+
+import nextstep.auth.authentication.AuthenticationToken;
+import nextstep.auth.context.Authentication;
+import nextstep.auth.context.SecurityContextHolder;
+import nextstep.auth.exception.AuthenticationException;
+import nextstep.member.application.LoginMemberService;
+import nextstep.member.domain.LoginMember;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public class UsernamePasswordAuthenticationFilter extends AuthenticationNonChainHandler {
+
+    public static final String USERNAME_FIELD = "username";
+    public static final String PASSWORD_FIELD = "password";
+
+    private final LoginMemberService loginMemberService;
+
+    public UsernamePasswordAuthenticationFilter(LoginMemberService loginMemberService) {
+        this.loginMemberService = loginMemberService;
+    }
+
+    @Override
+    protected LoginMember createAuthentication(HttpServletRequest request) {
+        String username = request.getParameter(USERNAME_FIELD);
+        String password = request.getParameter(PASSWORD_FIELD);
+
+        AuthenticationToken token = new AuthenticationToken(username, password);
+
+        LoginMember loginMember = loginMemberService.loadUserByUsername(token.getPrincipal());
+
+        if (loginMember == null) {
+            throw new AuthenticationException();
+        }
+
+        if (!loginMember.checkPassword(token.getCredentials())) {
+            throw new AuthenticationException();
+        }
+        return loginMember;
+    }
+
+    @Override
+    protected void afterHandle(LoginMember loginMember, HttpServletResponse response) {
+        Authentication authentication = new Authentication(loginMember.getEmail(), loginMember.getAuthorities());
+
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+    }
+}
